[Refactoring] moved-out file utils
